third_party/blink/web_tests/external/wpt/css/css-properties-values-api/animation/custom-property-animation-length-expected.txt

This is a testharness.js-based test.
PASS Animating a custom property of type <length>
PASS Animating a custom property of type <length> with a single keyframe
PASS Animating a custom property of type <length> with additivity
PASS Animating a custom property of type <length> with a single keyframe and additivity
FAIL Animating a custom property of type <length> with iterationComposite assert_equals: expected "250px" but got "50px"
Harness: the test ran to completion.
